refactor: 바텀시트 provider 상단으로 분리
