+++ app/(form)/_layout.tsx
@@ -1,8 +1,13 @@
 import React from 'react';
 import { Stack } from 'expo-router';
+import { BottomSheetModalProvider } from '@gorhom/bottom-sheet';
 
-function TabLayout() {
-  return <Stack screenOptions={{ headerShown: false }} />;
+function FormLayout() {
+  return (
+    <BottomSheetModalProvider>
+      <Stack screenOptions={{ headerShown: false }} />
+    </BottomSheetModalProvider>
+  );
 }
 
-export default TabLayout;
+export default FormLayout;
