+++ components/molecules/BottomSheetBackup/index.tsx
@@ -1,0 +1,41 @@
+import React, { useCallback, useMemo, useRef } from 'react';
+import { View, Text, StyleSheet } from 'react-native';
+import { BottomSheetBackdrop, BottomSheetModal } from '@gorhom/bottom-sheet';
+import IconButton from '@/components/molecules/IconButton';
+
+const App = () => {
+  const bottomSheetModalRef = useRef<BottomSheetModal>(null);
+  const snapPoints = useMemo(() => ['65%'], []);
+
+  const handlePresentModalPress = useCallback(() => {
+    bottomSheetModalRef.current?.present();
+  }, []);
+  const renderBackdrop = useCallback(
+    (props: any) => <BottomSheetBackdrop {...props} pressBehavior="close" appearsOnIndex={0} disappearsOnIndex={-1} />,
+    [],
+  );
+
+  return (
+    <View style={styles.container}>
+      <IconButton name="controller" onPress={handlePresentModalPress} />
+
+      <BottomSheetModal ref={bottomSheetModalRef} index={0} snapPoints={snapPoints} backdropComponent={renderBackdrop}>
+        <View style={styles.contentContainer}>
+          <Text>상세 검색창 아직 디자인 작업중이라고 함</Text>
+        </View>
+      </BottomSheetModal>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    justifyContent: 'center',
+  },
+  contentContainer: {
+    flex: 1,
+    alignItems: 'center',
+  },
+});
+
+export default App;
